Video_Generation_Engine/requirements.txt
pydantic>=2.0
PyYAML
python-dotenv
requests
